backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.3.0
dateparser==1.2.2
fastapi==0.118.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.6.3
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
